+++ modules/widgets/src/index.ts
@@ -1,0 +1,3 @@
+export {FullscreenWidget} from './fullscreen-widget';
+
+export * from './themes';